fix: remove requiring of missing "custom.utils"
The plugin is crashing on load due to the attempt to require
`custom.utils`. Removing the lines results in the plugin loading.

Considering that the requires are either assigned to an underscore
variable, or assigned to `utils` and then overwritten right after, I'm
guessing that these lines are OK to remove.